Moved tools and settings into their own folders
