+++ fast64_internal/sm64/settings/__init__.py
@@ -1,0 +1,8 @@
+from .panels import (
+    settings_panels_register,
+    settings_panels_unregister,
+)
+from .properties import (
+    settings_props_register,
+    settings_props_unregister,
+)
