+++ fast64_internal/sm64/settings/constants.py
@@ -1,0 +1,39 @@
+sm64GoalTypeEnum = [
+    ("All", "All", "All"),
+    ("Export Object/Actor/Anim", "Export Object/Actor/Anim", "Export Object/Actor/Anim"),
+    ("Export Level", "Export Level", "Export Level"),
+    ("Export Displaylist", "Export Displaylist", "Export Displaylist"),
+    ("Export UI Image", "Export UI Image", "Export UI Image"),
+]
+
+enumExportType = [
+    ("C", "C", "C"),
+    ("Binary", "Binary", "Binary"),
+    ("Insertable Binary", "Insertable Binary", "Insertable Binary"),
+]
+
+enumCompressionFormat = [
+    ("mio0", "MIO0", "MIO0"),
+    ("yay0", "YAY0", "YAY0"),
+]
+
+enumRefreshVer = [
+    ("Refresh 3", "Refresh 3", "Refresh 3"),
+    ("Refresh 4", "Refresh 4", "Refresh 4"),
+    ("Refresh 5", "Refresh 5", "Refresh 5"),
+    ("Refresh 6", "Refresh 6", "Refresh 6"),
+    ("Refresh 7", "Refresh 7", "Refresh 7"),
+    ("Refresh 8", "Refresh 8", "Refresh 8"),
+    ("Refresh 10", "Refresh 10", "Refresh 10"),
+    ("Refresh 11", "Refresh 11", "Refresh 11"),
+    ("Refresh 12", "Refresh 12", "Refresh 12"),
+    ("Refresh 13", "Refresh 13", "Refresh 13"),
+]
+
+draw_layer_enums = [
+    ("1", "Solid", "0x01"),
+    ("2", "Solid Decal", "0x02"),
+    ("4", "Transparent (No Blending)", "0x04"),
+    ("5", "Transparent (Blending Front)", "0x05"),
+    ("6", "Transparent (Blending Back)", "0x06"),
+]
