+++ fast64_internal/sm64/tools/__init__.py
@@ -1,0 +1,8 @@
+from .panels import (
+    tools_panels_register,
+    tools_panels_unregister,
+)
+from .operators import (
+    tools_operators_register,
+    tools_operators_unregister,
+)
